Add files via upload
Customer satisfaction has emerged as one of the most important factors 
that guarantee the success of online store; it has been posited as a key 
stimulant of purchase, repurchase intentions and customer loyalty. A 
comprehensive review of the literature, theories and models have been 
carried out to propose the models for customer activation and customer 
retention. Five major factors that contributed to the success of an eClassification: Internal
commerce store have been identified as: service quality, system quality, 
information quality, trust and net benefit. The research furthermore 
investigated the factors that influence the online customers repeat 
purchase intention. The combination of both utilitarian value and 
hedonistic values are needed to affect the repeat purchase intention 
(loyalty) positively. The data is collected from the Indian online shoppers. 
Results indicate the e-retail success factors, which are very much criticalCustomer satisfaction has emerged as one of the most important factors 
that guarantee the success of online store; it has been posited as a key 
stimulant of purchase, repurchase intentions and customer loyalty. A 
comprehensive review of the literature, theories and models have been 
carried out to propose the models for customer activation and customer 
retention. Five major factors that contributed to the success of an eClassification: Internal
commerce store have been identified as: service quality, system quality, 
information quality, trust and net benefit. The research furthermore 
investigated the factors that influence the online customers repeat 
purchase intention. The combination of both utilitarian value and 
hedonistic values are needed to affect the repeat purchase intention 
(loyalty) positively. The data is collected from the Indian online shoppers. 
Results indicate the e-retail success factors, which are very much critical for customer satisfaction.